# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "A simple E-Commerce API with Products and Orders"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/printer/orders/{id}/print": {
      "post": {
        "summary": "Print an order to ESC/POS printer",
        "tags": [
          "Printer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Print job sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "printed",
                        "devicePath",
                        "orderId"
                      ],
                      "properties": {
                        "printed": {
                          "type": "boolean"
                        },
                        "devicePath": {
                          "type": "string"
                        },
                        "orderId": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "printed": true,
                    "devicePath": "/dev/usb/lp0",
                    "orderId": 10
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "message": "Order not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}